Patch issue with bounces.
Reference: http://sentry.dev.lahavane.com/caraibes/mercurio/issues/2109/

It seems that (contrary to my current understanding of the addon) I'm
getting the old-plain tuple instead of the BounceVirtualId.

This commit only patches the issues instead of the actually solving the
problem.  Yet, I issue a warning to remind me.