Don't require alias' domain unless we set one.
If the alias does not have a domain set, don't require it.  Beware, that you
SHOULD configure your aliases with this feature.  Not doing it, implies that
if you get an email for your server with addresses for other domains, the
aliases may match (in Cc, for instance).

This fix an issue with test of the CRM addon.  They don't set the default
alias domain and send mails to the wrong domain anyways.