Don't send messages comming from Evaneos to Evaneos.
Evaneos intercepts all communications between us and the clients.  The emails
from the client reach our server From an Evaneos address.  Also it's T-Care
services uses an Evaneos address.

For a single dossier they use a unique identifier in the addresses.  And send
all the emails to a single alias.  We should not create an object each time
they send an email.  We locate the proper object and place the message there.

However, they're using at least two addresses for the client:

   traveler_XXXXX@m.evaneos.com
   traveler_XXXXX-UUID@m.evaneos.com

Which means that if we created an object and subscribe one address, all emails
coming from one the client's address are sent to its other address.  Similarly
for emails sent by Evaneos T-Care service.

The proposed solution is NOT to send messages coming from Evaneos to Evaneos.

The commit changes several things, because previously we disregard the names
in the emails, but now we need to send the recipients names.